+++ lib/simple_app_state.dart
@@ -1,3 +1,4 @@
-export 'src/core/simple_app_state_core.dart';
+export 'src/core/app_state_impl.dart'
+    show SimpleAppState, RefAppState, DebugListener, StateListener;
 export 'src/flutter/slot_stateful_widget.dart';
 export 'src/flutter/state_slot_builder.dart';