+++ lib/src/core/app_state_impl.dart
@@ -1,0 +1,7 @@
+import 'package:file_state_manager/file_state_manager.dart';
+import 'package:simple_app_state/src/core/app_state_protocol.dart';
+import 'package:simple_app_state/src/core/state_slot.dart';
+import 'package:simple_app_state/src/core/util_copy.dart';
+part 'simple_app_state_core.dart';
+part 'ref_app_state.dart';
+part 'app_state_internals.dart';